+++ samples/atta_vdisk.py
@@ -26,54 +26,39 @@
 # https://github.com/microsoft/Windows-classic-samples/blob/main/Samples/Hyper-V/Storage/cpp/AttachVirtualDisk.cpp
 
 import sys
-from ctwin32 import (
-    ATTACH_VIRTUAL_DISK_FLAG_PERMANENT_LIFETIME,
-    ATTACH_VIRTUAL_DISK_FLAG_READ_ONLY,
-    OPEN_VIRTUAL_DISK_VERSION_1,
-    OPEN_VIRTUAL_DISK_VERSION_2,
-    OPEN_VIRTUAL_DISK_FLAG_NONE,
-    VIRTUAL_DISK_ACCESS_READ,
-    VIRTUAL_DISK_ACCESS_NONE,
-    )
-from ctwin32.virtdisk import (
-    OPEN_VIRTUAL_DISK_PARAMETERS,
-    VIRTUAL_STORAGE_TYPE,
-    OpenVirtualDisk,
-    AttachVirtualDisk,
-    DetachVirtualDisk
-    )
+from ctwin32 import virtdisk as vdsk
 
 path = sys.argv[1]
 do_detach = len(sys.argv) > 2
 
-oparams = OPEN_VIRTUAL_DISK_PARAMETERS()
-oflags = OPEN_VIRTUAL_DISK_FLAG_NONE
+oparams = vdsk.OPEN_VIRTUAL_DISK_PARAMETERS()
+oflags = vdsk.OPEN_VIRTUAL_DISK_FLAG_NONE
 
 # Specify UNKNOWN for both device and vendor so the system will use
 # the file extension to determine the correct VHD format. The default values
 # do exactly that.
-storage_type = VIRTUAL_STORAGE_TYPE()
+storage_type = vdsk.VIRTUAL_STORAGE_TYPE()
 
 
 ext = path.rsplit(".", 1)[1].lower()
 if ext == "iso":
-    oparams.Version = OPEN_VIRTUAL_DISK_VERSION_1
-    acc_mask = VIRTUAL_DISK_ACCESS_READ
+    oparams.Version = vdsk.OPEN_VIRTUAL_DISK_VERSION_1
+    acc_mask = vdsk.VIRTUAL_DISK_ACCESS_READ
     aflags = (
-        ATTACH_VIRTUAL_DISK_FLAG_PERMANENT_LIFETIME
-        | ATTACH_VIRTUAL_DISK_FLAG_READ_ONLY
+        vdsk.ATTACH_VIRTUAL_DISK_FLAG_PERMANENT_LIFETIME
+        | vdsk.ATTACH_VIRTUAL_DISK_FLAG_READ_ONLY
         )
 else:
-    oparams.Version = OPEN_VIRTUAL_DISK_VERSION_2
+    oparams.Version = vdsk.OPEN_VIRTUAL_DISK_VERSION_2
     oparams.Version2.GetInfoOnly = False
-    aflags = ATTACH_VIRTUAL_DISK_FLAG_PERMANENT_LIFETIME
+    aflags = vdsk.ATTACH_VIRTUAL_DISK_FLAG_PERMANENT_LIFETIME
 
     # not tested -> CHANGE THIS
-    acc_mask = VIRTUAL_DISK_ACCESS_NONE
+    acc_mask = vdsk.VIRTUAL_DISK_ACCESS_NONE
 
 
-with OpenVirtualDisk(storage_type, path, acc_mask, oflags, oparams) as vd:
+with vdsk.OpenVirtualDisk(storage_type, path, acc_mask, oflags, oparams) as vd:
     if do_detach:
-        DetachVirtualDisk(vd)
+        vdsk.DetachVirtualDisk(vd)
     else:
-        AttachVirtualDisk(vd, aflags)
+        vdsk.AttachVirtualDisk(vd, aflags)