+++ changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   corresponding platform tag
 - fixed wrong \_raise_failed_status import in misc
 - fixed and simplified wtypes.ScdToBeClosed.from_param (no more integer overflow)
+- moved virtual disk constants from \_\_init\_\_.py to virtdisk.py
 
 ## [1.6.0]